backend/requirements.txt
Flask
Flask-Cors
pandas
scikit-learn
requests
python-dotenv
